[schemas][l]: schemas and test for it moved to one place and instantiate there - refs #376
